Vol/Nb blacklist Threshold in long instead of int

+++ openwis-dataservice/openwis-dataservice-server/openwis-dataservice-server-ejb/src/main/java/org/openwis/datasource/server/service/impl/BlacklistServiceImpl.java
@@ -310,9 +310,9 @@ public class BlacklistServiceImpl implements BlacklistService {
          result.setNbDisseminationWarnThreshold(JndiUtils
                .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_WARN));
          result.setNbDisseminationBlacklistThreshold(JndiUtils
-               .getInt(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_BLACKLIST));
+               .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_BLACKLIST));
          result.setVolDisseminationWarnThreshold(JndiUtils
-               .getInt(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_WARN));
+               .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_WARN));
          result.setVolDisseminationBlacklistThreshold(JndiUtils
                .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_BLACKLIST));
